back/news/requirements.txt
asgiref==3.3.0
certifi==2020.6.20
chardet==3.0.4
Django==3.1.3
django-cors-headers==3.5.0
django-google-auth==0.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
httplib2==0.18.1
idna==2.10
oauth2client==4.1.3
Pillow==8.0.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pytz==2020.4
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
